refactor: extract large modules into focused single-responsibility files
Split monolithic files for maintainability:
- platform.ts → platform/ (8 modules: git-ops, spawn, shell, dialogs, etc.)
- IssueManager → issue-manager-worktree, frontmatter, worktree-constants
- DashboardWriter → dashboard-writer-sort
- dashboard-issue-actions → dashboard-issue-color-dropdown
- IssueHeader → HeaderBadges, WorktreeIndicator subcomponents
- GitHubSearchContent → SearchButtonBar, SearchResultItem, config/engine/data-loaders
- git-status → git-badge-maps, git-status-helpers
