--- src/utils/platform.ts
@@ -1,901 +1,0 @@
-import { Notice, Platform } from 'obsidian';
-
-interface ShellApi {
-	openPath: (targetPath: string) => Promise<string>;
-}
-
-interface ChildProcessApi {
-	on: (event: 'error', listener: () => void) => ChildProcessApi;
-	unref: () => void;
-}
-
-interface SpawnApi {
-	(command: string, args?: string[], options?: Record<string, unknown>): ChildProcessApi;
-}
-
-interface FileDialogFilter {
-	name: string;
-	extensions: string[];
-}
-
-interface FileDialogResult {
-	canceled: boolean;
-	filePaths: string[];
-}
-
-interface RemoteDialogApi {
-	showOpenDialog: (options: {
-		properties: string[];
-		defaultPath?: string;
-		filters?: FileDialogFilter[];
-	}) => Promise<FileDialogResult>;
-}
-
-export interface WorktreeEntry {
-	path: string;
-	branch: string | undefined;
-	isBare: boolean;
-}
-
-export interface PlatformService {
-	openInFileExplorer: (folderPath: string) => void;
-	openTerminal: (folderPath: string, tabColor?: string) => void;
-	openVSCode: (folderPath: string, issueColor?: string) => void;
-	isGitRepositoryFolder: (folderPath: string) => boolean;
-	isGitBranchMissing: (folderPath: string, branchName: string) => boolean;
-	pathExists: (targetPath: string) => boolean;
-	findWorktreePathForBranch: (repositoryFolder: string, branchName: string) => string | undefined;
-	listActiveWorktrees: (repositoryFolder: string) => WorktreeEntry[];
-	getDefaultBranch: (repositoryFolder: string) => string | undefined;
-	getCurrentBranch: (repositoryFolder: string) => string | undefined;
-	checkBranchExists: (
-		repositoryFolder: string,
-		branchName: string
-	) => 'local' | 'remote' | 'none';
-	hasBranchUpstreamConfig: (repositoryFolder: string, branchName: string) => boolean;
-	pickFolder: (defaultPath?: string) => Promise<string | undefined>;
-	pickFile: (filters?: FileDialogFilter[], defaultPath?: string) => Promise<string | undefined>;
-	runWorktreeSetupScript: (
-		issueId: string,
-		color?: string,
-		dashboardWorkingDirectory?: string,
-		bashExecutablePath?: string
-	) => void;
-	runWorktreeRemovalScript: (
-		issueId: string,
-		dashboardWorkingDirectory?: string,
-		bashExecutablePath?: string,
-		options?: {
-			skipConfirmation?: boolean;
-			tabColor?: string;
-		}
-	) => boolean;
-}
-
-export interface ScriptPathResolver {
-	resolvePluginScriptPath: (filename: string) => string;
-}
-
-export function createPlatformService(scriptPathResolver?: ScriptPathResolver): PlatformService {
-	const HEX_COLOR_PATTERN = /^#[0-9A-Fa-f]{6}$/;
-	const WINDOWS_GIT_BASH_CANDIDATES = [
-		'C:\\_MP_apps\\Git\\bin\\bash.exe',
-		'C:\\Program Files\\Git\\bin\\bash.exe',
-		'C:\\Program Files (x86)\\Git\\bin\\bash.exe',
-		'C:\\_MP_apps\\Git\\git-bash.exe',
-		'C:\\Program Files\\Git\\git-bash.exe',
-		'C:\\Program Files (x86)\\Git\\git-bash.exe'
-	];
-	const isObjectRecord = (value: unknown): value is Record<string, unknown> => {
-		return typeof value === 'object' && value !== null;
-	};
-
-	const loadModule = (moduleName: string): unknown => {
-		// eslint-disable-next-line @typescript-eslint/no-require-imports
-		const loadedModule: unknown = require(moduleName);
-		return loadedModule;
-	};
-
-	const getRequiredFunction = (
-		moduleValue: unknown,
-		functionName: string
-	): ((...args: unknown[]) => unknown) => {
-		if (!isObjectRecord(moduleValue)) {
-			throw new Error('Invalid module value');
-		}
-
-		const candidateFunction = moduleValue[functionName];
-		if (typeof candidateFunction !== 'function') {
-			throw new Error(`Module function not found: ${functionName}`);
-		}
-
-		return (...args: unknown[]) => {
-			const result: unknown = Reflect.apply(candidateFunction, moduleValue, args);
-			return result;
-		};
-	};
-
-	const isPromiseString = (value: unknown): value is Promise<string> => {
-		if (!isObjectRecord(value)) {
-			return false;
-		}
-
-		return typeof value.then === 'function';
-	};
-
-	const getShell = (): ShellApi => {
-		const electronModule = loadModule('electron');
-		if (!isObjectRecord(electronModule)) {
-			throw new Error('Invalid electron module');
-		}
-
-		const shellValue = electronModule.shell;
-		if (!isObjectRecord(shellValue)) {
-			throw new Error('Invalid electron shell');
-		}
-
-		const openPathFunction = getRequiredFunction(shellValue, 'openPath');
-		return {
-			openPath: (targetPath: string) => {
-				const openPathResult = openPathFunction(targetPath);
-				if (isPromiseString(openPathResult)) {
-					return openPathResult;
-				}
-
-				return Promise.resolve('');
-			}
-		};
-	};
-
-	const toChildProcessApi = (processCandidate: unknown): ChildProcessApi => {
-		if (!isObjectRecord(processCandidate)) {
-			throw new Error('Invalid child process');
-		}
-
-		const onCandidate = processCandidate.on;
-		const unrefCandidate = processCandidate.unref;
-		if (typeof onCandidate !== 'function' || typeof unrefCandidate !== 'function') {
-			throw new Error('Invalid child process methods');
-		}
-
-		return {
-			on: (event: 'error', listener: () => void) => {
-				// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-				const onResult = onCandidate.call(processCandidate, event, listener);
-				return toChildProcessApi(onResult);
-			},
-			unref: () => {
-				unrefCandidate.call(processCandidate);
-			}
-		};
-	};
-
-	const getSpawn = (): SpawnApi => {
-		const childProcessModule = loadModule('child_process');
-		const spawnFunction = getRequiredFunction(childProcessModule, 'spawn');
-
-		return (
-			command: string,
-			args?: string[],
-			options?: Record<string, unknown>
-		): ChildProcessApi => {
-			const spawnedProcess = spawnFunction(command, args, options);
-			return toChildProcessApi(spawnedProcess);
-		};
-	};
-
-	interface SpawnSyncResult {
-		status: number | undefined;
-		stdout: string | undefined;
-	}
-
-	const runSpawnSync = (
-		folderPath: string,
-		args: string[],
-		options?: { encoding?: string }
-	): SpawnSyncResult => {
-		try {
-			const childProcessModule = loadModule('child_process');
-			const spawnSyncFunction = getRequiredFunction(childProcessModule, 'spawnSync');
-			const spawnResult = spawnSyncFunction('git', args, {
-				shell: false,
-				cwd: folderPath,
-				windowsHide: true,
-				...options
-			});
-			if (!isObjectRecord(spawnResult)) {
-				return { status: undefined, stdout: undefined };
-			}
-
-			const statusCandidate = spawnResult.status;
-			const status = typeof statusCandidate === 'number' ? statusCandidate : undefined;
-			const stdoutCandidate = spawnResult.stdout;
-			const stdout = typeof stdoutCandidate === 'string' ? stdoutCandidate : undefined;
-			return { status, stdout };
-		} catch {
-			return { status: undefined, stdout: undefined };
-		}
-	};
-
-	const runGitCommandStatus = (folderPath: string, args: string[]): number | undefined => {
-		return runSpawnSync(folderPath, args).status;
-	};
-
-	const runGitCommandOutput = (folderPath: string, args: string[]): string | undefined => {
-		const result = runSpawnSync(folderPath, args, { encoding: 'utf8' });
-		if (result.status !== 0) {
-			return undefined;
-		}
-		return result.stdout;
-	};
-
-	const isFileDialogResult = (value: unknown): value is FileDialogResult => {
-		if (!isObjectRecord(value)) {
-			return false;
-		}
-
-		const canceled = value.canceled;
-		const filePaths = value.filePaths;
-		if (typeof canceled !== 'boolean' || !Array.isArray(filePaths)) {
-			return false;
-		}
-
-		for (const filePath of filePaths) {
-			if (typeof filePath !== 'string') {
-				return false;
-			}
-		}
-
-		return true;
-	};
-
-	const getRemoteDialog = (): RemoteDialogApi => {
-		const remoteModule = loadModule('@electron/remote');
-		if (!isObjectRecord(remoteModule)) {
-			throw new Error('Invalid remote module');
-		}
-
-		const dialogCandidate = remoteModule.dialog;
-		if (!isObjectRecord(dialogCandidate)) {
-			throw new Error('Invalid remote dialog module');
-		}
-
-		const showOpenDialogFunction = getRequiredFunction(dialogCandidate, 'showOpenDialog');
-		return {
-			showOpenDialog: async (options: {
-				properties: string[];
-				defaultPath?: string;
-				filters?: FileDialogFilter[];
-			}): Promise<FileDialogResult> => {
-				const dialogResult = await showOpenDialogFunction(options);
-				if (!isFileDialogResult(dialogResult)) {
-					throw new Error('Invalid file dialog result');
-				}
-
-				return dialogResult;
-			}
-		};
-	};
-
-	const notifyOnSpawnError = (childProcess: ChildProcessApi, message: string): void => {
-		childProcess.on('error', () => {
-			new Notice(message);
-		});
-	};
-
-	const openInFileExplorer = (folderPath: string): void => {
-		void getShell().openPath(folderPath);
-	};
-
-	const isValidHexColor = (color: string): boolean => HEX_COLOR_PATTERN.test(color);
-
-	const openTerminal = (folderPath: string, tabColor?: string): void => {
-		const spawn = getSpawn();
-
-		if (Platform.isWin) {
-			const spawnArguments = ['-w', '0', 'nt', '-d', folderPath];
-			if (tabColor !== undefined && isValidHexColor(tabColor)) {
-				spawnArguments.push('--tabColor', tabColor);
-			}
-			notifyOnSpawnError(
-				spawn('wt', spawnArguments, { shell: false, cwd: folderPath }),
-				'Could not open Windows Terminal — is it installed?'
-			);
-		} else if (Platform.isMacOS) {
-			notifyOnSpawnError(
-				spawn('open', ['-a', 'Terminal', folderPath], { shell: false }),
-				'Could not open Terminal'
-			);
-		} else {
-			const terminalProcess = spawn(
-				'x-terminal-emulator',
-				['--working-directory', folderPath],
-				{ shell: false }
-			);
-			terminalProcess.on('error', () => {
-				// Fallback to xterm if x-terminal-emulator is not available
-				notifyOnSpawnError(
-					spawn('xterm', [], { shell: false, cwd: folderPath }),
-					'Could not open terminal — no terminal emulator found'
-				);
-			});
-		}
-	};
-
-	const ensurePeacockColor = (folderPath: string, issueColor: string): void => {
-		try {
-			const fsModule = loadModule('fs');
-			const pathModule = loadModule('path');
-			const joinFunction = getRequiredFunction(pathModule, 'join');
-			const existsSyncFunction = getRequiredFunction(fsModule, 'existsSync');
-			const readFileSyncFunction = getRequiredFunction(fsModule, 'readFileSync');
-			const writeFileSyncFunction = getRequiredFunction(fsModule, 'writeFileSync');
-			const mkdirSyncFunction = getRequiredFunction(fsModule, 'mkdirSync');
-
-			const vscodeDir = joinFunction(folderPath, '.vscode') as string;
-			const settingsPath = joinFunction(vscodeDir, 'settings.json') as string;
-
-			let settings: Record<string, unknown> = {};
-			if (existsSyncFunction(settingsPath) === true) {
-				const content = readFileSyncFunction(settingsPath, 'utf8') as string;
-				const parsed: unknown = JSON.parse(content);
-				if (typeof parsed === 'object' && parsed !== null) {
-					settings = parsed as Record<string, unknown>;
-				}
-			}
-
-			if (settings['peacock.color'] === issueColor) {
-				return;
-			}
-
-			settings['peacock.color'] = issueColor;
-			if (existsSyncFunction(vscodeDir) !== true) {
-				mkdirSyncFunction(vscodeDir, { recursive: true });
-			}
-			writeFileSyncFunction(settingsPath, JSON.stringify(settings, null, '\t') + '\n');
-		} catch {
-			// Silent failure — don't corrupt user settings
-		}
-	};
-
-	const openVSCode = (folderPath: string, issueColor?: string): void => {
-		if (issueColor !== undefined && isValidHexColor(issueColor)) {
-			ensurePeacockColor(folderPath, issueColor);
-		}
-		const spawn = getSpawn();
-		// shell: true required on Windows — `code` is a .cmd batch wrapper, not an executable
-		const child = spawn('code', [folderPath], {
-			shell: Platform.isWin,
-			cwd: folderPath,
-			detached: true,
-			stdio: 'ignore'
-		});
-		child.unref();
-		notifyOnSpawnError(child, 'Could not open VS Code — is it installed and in PATH?');
-	};
-
-	const isGitRepositoryFolder = (folderPath: string): boolean => {
-		if (folderPath.trim() === '') {
-			return false;
-		}
-
-		try {
-			const fsModule = loadModule('fs');
-			const existsSyncFunction = getRequiredFunction(fsModule, 'existsSync');
-			const normalizedPath = folderPath.replace(/[\\/]+$/, '');
-			const gitPath = `${normalizedPath}/.git`;
-			// .git can be a directory (regular repo) or a file (worktree)
-			return existsSyncFunction(gitPath) === true;
-		} catch {
-			return false;
-		}
-	};
-
-	const listActiveWorktrees = (repositoryFolder: string): WorktreeEntry[] => {
-		if (repositoryFolder.trim() === '' || !isGitRepositoryFolder(repositoryFolder)) {
-			return [];
-		}
-
-		const output = runGitCommandOutput(repositoryFolder, ['worktree', 'list', '--porcelain']);
-		if (output === undefined || output.trim() === '') {
-			return [];
-		}
-
-		const entries: WorktreeEntry[] = [];
-		let currentPath: string | undefined;
-		let currentBranch: string | undefined;
-		let isBare = false;
-		const lines = output.split(/\r?\n/);
-
-		for (const line of lines) {
-			const trimmedLine = line.trim();
-			if (trimmedLine === '') {
-				if (currentPath !== undefined) {
-					entries.push({ path: currentPath, branch: currentBranch, isBare });
-				}
-				currentPath = undefined;
-				currentBranch = undefined;
-				isBare = false;
-				continue;
-			}
-
-			if (trimmedLine.startsWith('worktree ')) {
-				currentPath = trimmedLine.slice('worktree '.length).trim();
-				continue;
-			}
-
-			if (trimmedLine.startsWith('branch refs/heads/')) {
-				currentBranch = trimmedLine.slice('branch refs/heads/'.length).trim();
-				continue;
-			}
-
-			if (trimmedLine === 'bare') {
-				isBare = true;
-			}
-		}
-
-		// Handle final entry without trailing empty line
-		if (currentPath !== undefined) {
-			entries.push({ path: currentPath, branch: currentBranch, isBare });
-		}
-
-		return entries;
-	};
-
-	const getDefaultBranch = (repositoryFolder: string): string | undefined => {
-		if (repositoryFolder.trim() === '' || !isGitRepositoryFolder(repositoryFolder)) {
-			return undefined;
-		}
-
-		// Try symbolic-ref for origin HEAD first
-		const symbolicRefOutput = runGitCommandOutput(repositoryFolder, [
-			'symbolic-ref',
-			'refs/remotes/origin/HEAD',
-			'--short'
-		]);
-		if (symbolicRefOutput !== undefined && symbolicRefOutput.trim() !== '') {
-			const trimmed = symbolicRefOutput.trim();
-			// Output is like "origin/main" — strip the "origin/" prefix
-			const slashIndex = trimmed.indexOf('/');
-			if (slashIndex !== -1) {
-				return trimmed.slice(slashIndex + 1);
-			}
-			return trimmed;
-		}
-
-		// Fall back to current HEAD branch
-		const headOutput = runGitCommandOutput(repositoryFolder, [
-			'rev-parse',
-			'--abbrev-ref',
-			'HEAD'
-		]);
-		if (headOutput !== undefined && headOutput.trim() !== '' && headOutput.trim() !== 'HEAD') {
-			return headOutput.trim();
-		}
-
-		return undefined;
-	};
-
-	const pathExists = (targetPath: string): boolean => {
-		if (targetPath.trim() === '') {
-			return false;
-		}
-
-		try {
-			const fsModule = loadModule('fs');
-			const existsSyncFunction = getRequiredFunction(fsModule, 'existsSync');
-			return existsSyncFunction(targetPath) === true;
-		} catch {
-			return false;
-		}
-	};
-
-	const findWorktreePathForBranch = (
-		repositoryFolder: string,
-		branchName: string
-	): string | undefined => {
-		if (repositoryFolder.trim() === '' || branchName.trim() === '') {
-			return undefined;
-		}
-		if (isUnsafeScriptArgument(branchName)) {
-			return undefined;
-		}
-		if (!isGitRepositoryFolder(repositoryFolder)) {
-			return undefined;
-		}
-
-		const output = runGitCommandOutput(repositoryFolder, ['worktree', 'list', '--porcelain']);
-		if (output === undefined || output.trim() === '') {
-			return undefined;
-		}
-
-		let currentWorktreePath: string | undefined;
-		let currentBranch: string | undefined;
-		const lines = output.split(/\r?\n/);
-		for (const line of lines) {
-			const trimmedLine = line.trim();
-			if (trimmedLine === '') {
-				if (currentBranch === branchName && currentWorktreePath !== undefined) {
-					return currentWorktreePath;
-				}
-				currentWorktreePath = undefined;
-				currentBranch = undefined;
-				continue;
-			}
-
-			if (trimmedLine.startsWith('worktree ')) {
-				currentWorktreePath = trimmedLine.slice('worktree '.length).trim();
-				continue;
-			}
-
-			if (trimmedLine.startsWith('branch refs/heads/')) {
-				currentBranch = trimmedLine.slice('branch refs/heads/'.length).trim();
-			}
-		}
-
-		if (currentBranch === branchName && currentWorktreePath !== undefined) {
-			return currentWorktreePath;
-		}
-
-		return undefined;
-	};
-
-	const isGitBranchMissing = (folderPath: string, branchName: string): boolean => {
-		if (folderPath.trim() === '' || branchName.trim() === '') {
-			return false;
-		}
-		if (isUnsafeScriptArgument(branchName)) {
-			return false;
-		}
-		if (!isGitRepositoryFolder(folderPath)) {
-			return false;
-		}
-
-		const gitVersionStatus = runGitCommandStatus(folderPath, ['--version']);
-		if (gitVersionStatus !== 0) {
-			return false;
-		}
-
-		const localBranchStatus = runGitCommandStatus(folderPath, [
-			'rev-parse',
-			'--verify',
-			'--quiet',
-			`refs/heads/${branchName}`
-		]);
-		if (localBranchStatus === 0) {
-			return false;
-		}
-
-		const remoteBranchStatus = runGitCommandStatus(folderPath, [
-			'rev-parse',
-			'--verify',
-			'--quiet',
-			`refs/remotes/origin/${branchName}`
-		]);
-		return remoteBranchStatus !== 0;
-	};
-
-	const pickFolder = async (defaultPath?: string): Promise<string | undefined> => {
-		try {
-			const remoteDialog = getRemoteDialog();
-			const result = await remoteDialog.showOpenDialog({
-				properties: ['openDirectory'],
-				defaultPath
-			});
-			if (result.canceled || result.filePaths.length === 0) {
-				return undefined;
-			}
-			return result.filePaths[0];
-		} catch {
-			new Notice('Could not open folder picker');
-			return undefined;
-		}
-	};
-
-	const pickFile = async (
-		filters?: FileDialogFilter[],
-		defaultPath?: string
-	): Promise<string | undefined> => {
-		try {
-			const remoteDialog = getRemoteDialog();
-			const result = await remoteDialog.showOpenDialog({
-				properties: ['openFile'],
-				defaultPath,
-				filters
-			});
-			if (result.canceled || result.filePaths.length === 0) {
-				return undefined;
-			}
-			return result.filePaths[0];
-		} catch {
-			new Notice('Could not open file picker');
-			return undefined;
-		}
-	};
-
-	const toWindowsBashPath = (pathValue: string): string => {
-		const normalizedPath = pathValue.replace(/\\/g, '/');
-		const windowsPathMatch = normalizedPath.match(/^([A-Za-z]):\/(.*)$/);
-		if (windowsPathMatch === null) {
-			return normalizedPath;
-		}
-
-		const driveLetter = windowsPathMatch[1].toLowerCase();
-		const remainingPath = windowsPathMatch[2];
-		return `/${driveLetter}/${remainingPath}`;
-	};
-
-	const toBashSingleQuoted = (value: string): string => {
-		return `'${value.replace(/'/g, `'\\''`)}'`;
-	};
-
-	const getScriptWorkingDirectory = (scriptPath: string): string => {
-		return scriptPath.replace(/\\scripts\\[^\\]+$/, '');
-	};
-
-	const resolveScriptWorkingDirectory = (
-		scriptPath: string,
-		dashboardWorkingDirectory?: string
-	): string => {
-		if (dashboardWorkingDirectory !== undefined && dashboardWorkingDirectory.trim() !== '') {
-			return dashboardWorkingDirectory;
-		}
-
-		return getScriptWorkingDirectory(scriptPath);
-	};
-
-	const resolveWindowsBashExecutable = (configuredPath?: string): string | undefined => {
-		if (!Platform.isWin) {
-			return undefined;
-		}
-
-		const fsModule = loadModule('fs');
-		const existsSyncFunction = getRequiredFunction(fsModule, 'existsSync');
-		if (configuredPath !== undefined && configuredPath.trim() !== '') {
-			return configuredPath;
-		}
-
-		const bashCandidates = WINDOWS_GIT_BASH_CANDIDATES;
-		for (const bashCandidate of bashCandidates) {
-			const existsResult = existsSyncFunction(bashCandidate);
-			if (existsResult === true) {
-				return bashCandidate;
-			}
-		}
-
-		return undefined;
-	};
-
-	const UNSAFE_SCRIPT_ARGUMENT_PATTERN = /[\0\n\r`$|&;><(){}]/;
-	const isUnsafeScriptArgument = (value: string): boolean => {
-		return UNSAFE_SCRIPT_ARGUMENT_PATTERN.test(value);
-	};
-
-	const buildBashCommand = (scriptPath: string, args: string[]): string => {
-		const quotedScriptPath = toBashSingleQuoted(toWindowsBashPath(scriptPath));
-		const quotedArgs = args.map((arg) => toBashSingleQuoted(arg));
-		return ['bash', quotedScriptPath, ...quotedArgs].join(' ');
-	};
-
-	const buildInteractiveWindowsScriptCommand = (scriptPath: string, args: string[]): string => {
-		const scriptCommand = buildBashCommand(scriptPath, args);
-		return `(${scriptCommand} && echo setup_worktree_completed || echo setup_worktree_failed_exit_${'$'}?) && bash -i || bash -i`;
-	};
-
-	const runScriptWithBash = (
-		scriptPath: string,
-		args: string[],
-		errorMessage: string,
-		dashboardWorkingDirectory?: string,
-		bashExecutablePath?: string,
-		terminalTabColor?: string
-	): boolean => {
-		const spawn = getSpawn();
-		const workingDirectory = resolveScriptWorkingDirectory(
-			scriptPath,
-			dashboardWorkingDirectory
-		);
-		if (Platform.isWin) {
-			const bashExecutable = resolveWindowsBashExecutable(bashExecutablePath);
-			if (bashExecutable === undefined) {
-				new Notice(
-					'Bash executable was not found. Install/configure bash to avoid fallback.'
-				);
-				return false;
-			}
-
-			const isGitBashTerminal = bashExecutable.toLowerCase().endsWith('git-bash.exe');
-			const bashArgs = isGitBashTerminal
-				? [
-						`--cd=${workingDirectory}`,
-						'-i',
-						'-l',
-						'-c',
-						buildInteractiveWindowsScriptCommand(scriptPath, args)
-					]
-				: ['-i', '-l', '-c', buildInteractiveWindowsScriptCommand(scriptPath, args)];
-			const windowsTerminalArgs = ['-w', '0', 'nt', '-d', workingDirectory];
-			if (terminalTabColor !== undefined && isValidHexColor(terminalTabColor)) {
-				windowsTerminalArgs.push('--tabColor', terminalTabColor);
-			}
-			windowsTerminalArgs.push(bashExecutable, ...bashArgs);
-			const childProcess = spawn('wt', windowsTerminalArgs, {
-				shell: false,
-				cwd: workingDirectory
-			});
-			notifyOnSpawnError(childProcess, errorMessage);
-			return true;
-		}
-
-		const childProcess = spawn('bash', [scriptPath, ...args], {
-			shell: false,
-			cwd: workingDirectory
-		});
-		notifyOnSpawnError(childProcess, errorMessage);
-		return true;
-	};
-
-	const runWorktreeSetupScript = (
-		issueId: string,
-		color?: string,
-		dashboardWorkingDirectory?: string,
-		bashExecutablePath?: string
-	): void => {
-		if (isUnsafeScriptArgument(issueId)) {
-			new Notice('Invalid issue ID — contains forbidden characters.');
-			return;
-		}
-
-		if (scriptPathResolver === undefined) {
-			new Notice('Worktree scripts not available — plugin path could not be resolved.');
-			return;
-		}
-
-		const setupScriptPath = scriptPathResolver.resolvePluginScriptPath('setup-worktree.sh');
-
-		const scriptArgs = [issueId];
-		if (color !== undefined && color !== '' && isValidHexColor(color)) {
-			scriptArgs.push('--color', color);
-		}
-		runScriptWithBash(
-			setupScriptPath,
-			scriptArgs,
-			'Could not run setup-worktree script',
-			dashboardWorkingDirectory,
-			bashExecutablePath,
-			color
-		);
-	};
-
-	const runWorktreeRemovalScript = (
-		issueId: string,
-		dashboardWorkingDirectory?: string,
-		bashExecutablePath?: string,
-		options?: {
-			skipConfirmation?: boolean;
-			tabColor?: string;
-		}
-	): boolean => {
-		if (isUnsafeScriptArgument(issueId)) {
-			new Notice('Invalid issue ID — contains forbidden characters.');
-			return false;
-		}
-
-		if (scriptPathResolver === undefined) {
-			new Notice('Worktree scripts not available — plugin path could not be resolved.');
-			return false;
-		}
-
-		const removeScriptPath = scriptPathResolver.resolvePluginScriptPath('remove-worktree.sh');
-
-		const scriptArgs = [issueId];
-		if (options?.skipConfirmation === true) {
-			scriptArgs.push('--skip-confirmation');
-		}
-
-		return runScriptWithBash(
-			removeScriptPath,
-			scriptArgs,
-			'Could not run remove-worktree script',
-			dashboardWorkingDirectory,
-			bashExecutablePath,
-			options?.tabColor
-		);
-	};
-
-	const getCurrentBranch = (repositoryFolder: string): string | undefined => {
-		if (repositoryFolder.trim() === '' || !isGitRepositoryFolder(repositoryFolder)) {
-			return undefined;
-		}
-
-		const output = runGitCommandOutput(repositoryFolder, ['rev-parse', '--abbrev-ref', 'HEAD']);
-		if (output === undefined || output.trim() === '' || output.trim() === 'HEAD') {
-			return undefined;
-		}
-
-		return output.trim();
-	};
-
-	const checkBranchExists = (
-		repositoryFolder: string,
-		branchName: string
-	): 'local' | 'remote' | 'none' => {
-		if (repositoryFolder.trim() === '' || branchName.trim() === '') {
-			return 'none';
-		}
-		if (isUnsafeScriptArgument(branchName)) {
-			return 'none';
-		}
-		if (!isGitRepositoryFolder(repositoryFolder)) {
-			return 'none';
-		}
-
-		// Check remote first — a pushed branch is active regardless of local state
-		const remoteStatus = runGitCommandStatus(repositoryFolder, [
-			'rev-parse',
-			'--verify',
-			'--quiet',
-			`refs/remotes/origin/${branchName}`
-		]);
-		if (remoteStatus === 0) {
-			return 'remote';
-		}
-
-		const localStatus = runGitCommandStatus(repositoryFolder, [
-			'rev-parse',
-			'--verify',
-			'--quiet',
-			`refs/heads/${branchName}`
-		]);
-		if (localStatus === 0) {
-			return 'local';
-		}
-
-		// Fallback: git branch --list is more portable on Windows/MINGW64
-		const branchListOutput = runGitCommandOutput(repositoryFolder, [
-			'branch',
-			'--list',
-			branchName
-		]);
-		if (branchListOutput !== undefined && branchListOutput.trim() !== '') {
-			return 'local';
-		}
-
-		return 'none';
-	};
-
-	const hasBranchUpstreamConfig = (repositoryFolder: string, branchName: string): boolean => {
-		if (repositoryFolder.trim() === '' || branchName.trim() === '') {
-			return false;
-		}
-		if (isUnsafeScriptArgument(branchName)) {
-			return false;
-		}
-
-		// git config branch.<name>.remote — returns 0 if upstream tracking was configured
-		const status = runGitCommandStatus(repositoryFolder, [
-			'config',
-			`branch.${branchName}.remote`
-		]);
-		return status === 0;
-	};
-
-	return {
-		openInFileExplorer,
-		openTerminal,
-		openVSCode,
-		isGitRepositoryFolder,
-		isGitBranchMissing,
-		pathExists,
-		findWorktreePathForBranch,
-		listActiveWorktrees,
-		getDefaultBranch,
-		getCurrentBranch,
-		checkBranchExists,
-		hasBranchUpstreamConfig,
-		pickFolder,
-		pickFile,
-		runWorktreeSetupScript,
-		runWorktreeRemovalScript
-	};
-}
